modified

+++ SORTING/BucketSort.java
@@ -12,7 +12,7 @@ public class BucketSort {
             BucketSort.put(i,new LinkedList<>());
 
         for (int i=0; i< arr.length;i++)
-            BucketSort.get((arr[i]*nbksBuckets)).add(arr[i]);
+            BucketSort.get((int)(arr[i]*nbksBuckets)).add(arr[i]);
 
         for (int i=0;i< nbksBuckets;i++)
             Collections.sort(BucketSort.get(i));
